Basic authentification, basic security, added confirmation to delete a post

+++ templates/base/admin.tpl
@@ -4,8 +4,13 @@
         <tr class="arts">
             <td><a href="perm.php?id={$i.id}">{$i.title}</a></td>
             <td><a href="editor.php?id={$i.id}">Edit</a></td>
-            <td><a href="admin.php?delete={$i.id}">Delete</a></td>
+            <td><a href="admin.php?delete={$i.id}" class="confirmation">Delete</a></td>
         </tr>
     {/foreach}
 </table>
+    <script type="text/javascript">
+    $('.confirmation').on('click', function () {
+        return confirm('This post will be deleted, are you sure ?');
+    });
+</script>
 {include file='base/footer.tpl'}
